Continued breaking up the countTags() method to take a pair of files and simultaneously work through them.
Removed an top-level controlling loop for the read directions.  Instead
just hard-coded in 0 (forward) and 1 (reverse) since this convention
will be followed throughout all multidimensional arrays.

Using arrays for enzyme names and key files

Found out that the taxaNames array and loops that pull data into it
are not used anywhere downstream from its declaration and population.  I
assume that it was once intended to do something but was abandoned and
never removed.  For the time being I have just commented it out. 
Running the program with this variable commented out does not seem to
impact the final output.